+++ packages/backend-next/CHANGELOG.md
@@ -1,5 +1,12 @@
 # example-backend-next
 
+## 0.0.5-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-scaffolder-backend@1.8.0-next.1
+
 ## 0.0.5-next.0
 
 ### Patch Changes