+++ plugins/scaffolder-backend-module-yeoman/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @backstage/plugin-scaffolder-backend-module-yeoman
 
+## 0.2.11-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-scaffolder-backend@1.8.0-next.1
+
 ## 0.2.11-next.0
 
 ### Patch Changes